adding new endpoint for user lookup to bypass modsec rule
